redesigned broadcastController and userRegistrator - no any unnecessary channels
