+++ engine/userhandler.go
@@ -1,46 +1,113 @@
 package engine
 
 import (
+	"encoding/json"
 	"log"
-	"net/http"
+	"math/rand"
+	"sync"
 	"time"
 
 	"github.com/gorilla/websocket"
 )
 
-var upgrader = websocket.Upgrader{
-	ReadBufferSize:  1024,
-	WriteBufferSize: 1024,
-	CheckOrigin:     func(r *http.Request) bool { return true }, //  TODO need to develop a check that connection is from applicable client
+type UserData struct {
+	userID      int32
+	name        string
+	client      chan Message
+	connection  *websocket.Conn
+	chatID      string
+	broadcaster *broadcast
+	publicKey   string
 }
 
-const (
-	// Time allowed to write a message to the peer.
-	writeWait = 10 * time.Second
-)
+type Peers struct {
+	sync.RWMutex
+	peers map[int32]*UserData
+}
 
-func UserRegistrator(w http.ResponseWriter, r *http.Request) {
-	conn, err := upgrader.Upgrade(w, r, nil)
-	if err != nil {
-		log.Println(err)
-		return
+type Message struct {
+	MsgType string
+	Name    string
+	ChatID  string
+	Message string
+}
+
+func NewClient(conn *websocket.Conn, br *broadcast, name, chatID string) *UserData {
+	return &UserData{
+		userID:      rand.Int31(),
+		client:      make(chan Message),
+		connection:  conn,
+		name:        name,
+		chatID:      chatID,
+		broadcaster: br}
+}
+
+func (cli *UserData) SetPublicKey(key string) {
+	cli.publicKey = key
+}
+
+func (cli *UserData) clientReader() {
+	conn := cli.connection
+	br := cli.broadcaster
+	cli.client <- *cli.composeMessage("SRV", "Welcome, "+cli.name)
+	br.entering <- cli
+	br.messages <- *cli.composeMessage("SRV", cli.name+" joined the room")
+
+	for {
+		_, message, err := conn.ReadMessage()
+		if err != nil {
+			if websocket.IsUnexpectedCloseError(err, websocket.CloseGoingAway, websocket.CloseAbnormalClosure) {
+				log.Printf("error: %v", err)
+			}
+			break
+		}
+		var msg Message
+		if err := json.Unmarshal(message, &msg); err != nil {
+			log.Printf("%s - %s\n", err, "message JSON was not unmarshalled")
+		}
+		switch msg.MsgType {
+		case "MSG":
+			br.messages <- msg
+		case "KEY":
+			cli.SetPublicKey(msg.Message)
+			br.handshake <- cli
+		}
+
 	}
-	name, ok := r.URL.Query()["name"]
-	if !ok || len(name[0]) == 0 {
-		log.Println("Name was not send")
-		return
+
+	br.leaving <- cli
+	br.messages <- *cli.composeMessage("SRV", cli.name+" has left")
+	conn.Close()
+}
+
+func (cli *UserData) clientWriter() {
+	conn := cli.connection
+	ch := cli.client
+	for msg := range ch {
+		conn.SetWriteDeadline(time.Now().Add(writeWait))
+
+		w, err := conn.NextWriter(websocket.TextMessage)
+		if err != nil {
+			return
+		}
+		jsonMsg, err := json.Marshal(msg)
+		if err != nil {
+			log.Printf("%s - %s\n", err, "Message struct was not marshalled")
+		}
+		w.Write(jsonMsg)
+
+		for i := 0; i < len(ch); i++ {
+			w.Write([]byte{'\n'})
+			jsonMsg, _ := json.Marshal(<-ch)
+			w.Write(jsonMsg)
+		}
+
+		if err := w.Close(); err != nil {
+			return
+		}
 	}
-	chatID, ok := r.URL.Query()["chatID"]
-	if !ok || len(name[0]) == 0 {
-		log.Println("chatID was not send")
-		return
-	}
-	broadcastRequest <- chatID[0]
-	br, ok := <-broadcastAnswer
-	if !ok {
-		return
-	}
-	cli := NewClient(conn, br, name[0], chatID[0])
-	go cli.clientReader()
-	go cli.clientWriter()
+}
+
+func (cli *UserData) composeMessage(msgType string, message string) *Message {
+	return &Message{msgType, cli.name, cli.chatID, message}
 }
